исправил тип delay с float на int (в милисекундах)

+++ main.py
@@ -10,7 +10,7 @@ from gameController import gameController
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('-s', '--size', type=str, dest='size', default='50 21', help='Size of map, pair of integer as "x_size y_size". By default equals to "50 21"')
-    parser.add_argument('-d', '--delay', type=float, dest='delay', default=500, help='Delay between turns in ms.')
+    parser.add_argument('-d', '--delay', type=int, dest='delay', default=500, help='Delay between turns in ms.')
     parser.add_argument('--logs', action='store_true', dest='logs_flag', help='Enable logs gathering.')
     parser.add_argument('-t', '--theme', type=str, dest='theme', default='constructor', help='Theme for graphical interface.')
